Alas, it's the end of the road for the wallpaper apps! It turns out Apple Guideline 4.3 sees my apps as spam, which defeats the intention of the wallpaper apps being an income base which is genuinely allowed, rather than built on any shaky ground. I could technically keep submitting apps, and having them approved, but after this one was rejected, it would only be a matter of time until my account would be banned, from what I can tell, which would not be ideal at all.
And I'm not sure I'm willing to put any energy into something which doesn't have that more solid base of policy-abiding which I was looking for (i.e, sinking 10 hours into something only for it to vanish overnight). There's always risks in stuff, but it when if messes with the rules which allow one to operate, it seems unwise, as you don't want your income source to be removed out of nowhere, where you can't do anything to fight against is because it's literally in the rules).

And hence, alas!

It was a good run though, and I have nice example code and apps online for future job applications now. I'll probably do some more commits to main in the future just to do some cleanup (mainly just moving protocols and enums into their own files)